+++ injectable/example/bin/main.dart
@@ -1,5 +1,4 @@
 import 'package:example/injector/injector.dart';
-import 'package:example/module/register_module.dart';
 import 'package:get_it/get_it.dart';
 
 GetIt getIt = GetIt.instance;